Consumers shouldn't need TypeScript to build Polaris
Reason:
Previously, the only option for a consumer to treeshake Polaris was to add a TypeScript dependency.  This is bad because it:
  - Forces all consumers to update TypeScript in lockstep with Polaris
  - Causes type errors if a consumer doesn't peg their `@types` to specific versions (especially React)
  - Forces JS projects to install an unneeded dependency
  - Increases build times

Solution:
Create an `esnext` dir with:
  - JSX tranlated; no other babel transorms applied
  - Minified CSS

Note: this PR doesn't touch the build segment; I'm doing just enough to make this testable in `sewing-kit`/`rails_sewing_kit_example`.