Add publication filtering to resource picker
By default the resource picker will show all products and collections on
a shop, and visually denote which ones are "Hidden" in the case of
resources not published to a channel. This adds support for a boolean
flag named "showHidden" that defaults to true to preserve the current
behaviour.

See: https://github.com/Shopify/shopify/issues/132976