Stop prettier rampaging over markdown files
The 'proseWrap' setting forces line wrapping in markdown files, which
results in really blatant reformatting of everything. I'm not currently
convinced we want that as we may often want to rely on soft-wrapping as
Github-flavoured markdown is sometimes sensitive to newlines.

Go back to the default behaviour of prettier which is
`proseWrap:"preserve"`.

Remove explicitly setting the printWidth as it defaults to 80
anyway.