Make pre-commit hook pass if playground is not staged
Previously the pre-commit hook stopped you commiting if you had any
changes in the playground even if those changes weren't staged for
commit. This meant you had to totally reset your playground even if you
weren't about to commit it.

Now we check if the playground is staged for commit, so if the
playground is modified but not staged for commit you can still commit
your other changes