Enable shopify/no-ancestor-directory-import eslint rule
Importing from the component index leads to circular depdendencies that
cause issues when trying to optimize builds upstream. See this essay
https://github.com/Shopify/sewing-kit/blob/2bbf40fee4b6fc6e8a2cbbab1b743b47532b98eb/src/tools/webpack/tests/config/rules/javascript.test.ts#L215-L268

This commit adds tooling to help idenify cycles and stop them, so that
sewing-kit can remove its workarounds.

* Depdendency Cruiser (`yarn run depcruise-validate`) shall help
idenfity cycles.
* shopify/no-ancestor-directory-import is an eslint rule that bars
imports from parent indexes, which is the most common cause of cycles.

Fix current ancestort imports to import from direct files. This is not
enforced in tests because. test files will not be imported by anything
else so there is no danger of cylical imports.