Give Button size an explictly named default
You can now do `<Button size="medium"/>` to create a button with the same
size as `<Button>`. Prior to this commit there was no named size that
corresponded to the default of omiting the size attribute.

This is consistent with other components that can have a size set (e.g.
Avatar, Thumbnail and Spinner).