Move src/tests/utilities to src/test-utilities
Having both a /tests and /src/tests confuses the module lookup when we
try to make use of ts-jest to get sourcemaps working as ts-jest wants to
map 'test/utilities' to '/tests/utilities' rather than
'/src/tests/utilities'.

This is solvable by specifying a module resolution in the jest config,
but we can avoid that extra config and make things a bit less confusing
by avoiding the conflict by renaming the folder.