Fix consistency issue in prop types (#727)
 Fix the errors in the prop types that were being caught by the Props Explorer
These errors were happening because the explorer makes the assumption that all components export a "Props" interface that is then used by the component as the type of its props. When this is not the case, the typedoc output changes format, and the props explorer can't handle this and just silently fails.