Rename Option and Group types in Select to workaround typedoc oddness
Typedoc (which we use for generating prop tables on the styleguide) has
a bug in it where it doesn't correctly resolve types that are imported
from a different file (or we're using it wrong).
Instead of looking at imports it seems to look at a global bucket of
names. In this case there was a clash between Option (as living in Select)
and Option (as living in OptionList/components). Work around this by
renaming Select's Option to "SelectOption".

This allows our type docmentation generation in the styleguide to work
correctly.

Strictly speaking the change to SelectGroup isn't needed but i think
"SelectOption and SelectGroup" sounds nicer than "SelectOption and
Group".