Distribute compiled sass in the styles folder
This helps protect us against people overriding our variables / mixins /
functions and thus causing problematic output.

Do the sass build as part of the rollup styles plugin instead of a
separate step. This saves us having to generate a file in the rollup
plugin that we then pass on to a second step.

Merge src/styles/global into a single file, it's pretty small and it's
easier to deal with a single file