Styles build: combine all components into one file
There is no value in having individual files per component in the styles
folder. It is unneeded overhead and people shouldn't be digging into
that folder trying to include a single component's styles in the first place.

Where styles/components.scss previously was a list imports to
per-component files, instead inline the contents of the per-component scss
into styles/components.scss. The listing of the contents is preserved so
the computed css out the other end will be idential.

At this point global.scss and components.scss both contain pure css.
They are only distributed as scss files for the sake of convinience and
ease of @importing.