Fixup stylelint setup in VSCode after plugin changes (#2484)
The shinn.stylelint extension has been removed from the marketplace,
suggest using hex-ci.stylelint-plus instead as it is a an up-to-date
fork that adds autofixing.

Add autofixing config to vscode. This ensures rules autofixable
stylelint rules are fixed on save. This stopped working recently when
the prettier extension changed how it handled stylelint integration (it
now suggests using the autofixing built into the stylelint-plus and
eslint extensions directly)