Update storybook to use production babel config
Previously we used "not-production" because I thought the
babel-plugin-transform-react-constant-elements plugin messed things up,
but it turns out it was my complicated codeInvoker stuff in the readme
loader.

This helps prove that said plugin isn't messing up our builds.

This removes the envName config (as webpack will set it to "production"
in production mode automatically, and removes the minified option, as
webpack uses Terser in production mode to minify everything anyway so we
don't need Babel doing it too.

The readme loader now requires a manual list of polaris imports that
should be available (just like icons). Eventually it'd be good if
READMEs could have some metadata specifying their what imports they
require instead of us importing everything every time, but that's a plan
for later.