Nest split button example in tall enough <div>
The example currently does not render well when the button dropdown is
clicked, as it renders past the boundaries of the host div, and is
obscured. Due to this, it appears that the button is broken and does
nothing.

This follows the pattern in <Autocomplete />'s README and host the
example in a div with enough height.