Fix issue with dummy section being inserted
There was an issue taking place with extra markup making it's way into
the DOM when using this component.

This was due to this component inserting a "dummy" section into the
array of options being returned from normalizedOptions.

Basically we were inserting a dummy section with the component title and
empty options at the beginning of the normalizedOptions array.

When we mapped that array to Option components we would render a option
with a title and no options which isn't great semantically and if
someone put some styling into the `li` element this would really stand
out.