Unify the output of the styles folder build
Do a funky thing with how we reference a global animation to avoid the
usage of `:global` in our mixins. This means that contents of the mixins
for use by consumers who don't use css-modules (/styles) and those that
do use css-modules (/esnext/styles) are identical.

This allows us to remove some manipulation in the build now.
We can also tell all consumers to pull in the sass files via
`styles/_public-api.scss` instead of `esnext/styles/_public-api.scss`.

We can't fully delete `esnext/styles` just yet as the raw scss files in
the esnext folder still reference the contents of that folder.
Once we start shipping built instead of source scss files in the esnext
folder we would be able to remove esnext/styles entirely.