Use esnext as custom mainfield for esnext build
sewing-kit v0 now uses `esnext` as the custom mainField, to match the
behaviour of sewing-kit v1. Update to match that.

This is not a breaking change as sewing-kit v0 still has an alias forces
the use of the esnext folder, so this mainField doesn't do anything
currently, it's just for future compatability, once SK v0 drops support
for Polaris <5.0 - which will now be "dropping support for < 5.2".