Use contrasting colors in theming examples
If we only pass in a "background" to the legacy theme then we adjust its
lightness, however changing lightness does not guarantee in a large
enough contrast.

This problem goes away with new theming so no need to think too hard
about it.

For now we can make use of accessible colors in our examples to keep the
a11y tests happy.