Use prepublishOnly step instead of prepublish
npm has depcrated the prepublish lifecycle step, because running this
step as part of a publish and on install is confusing. And if you only
wanted something to run as part of a publish you had to gate it with an
`in-publish script`.

prepublishOnly is the replacement that runs a command only as part of
publish. Which means we don't need that `in-publish` check

See https://docs.npmjs.com/cli/v6/using-npm/scripts#prepare-and-prepublish