Use onChange if onClearButtonClick is not provided
If onClearButtonClick is not provided, fallback to calling onChange with
an empty string. This allows users to enable clearing of the field by
just adding `clearButton={true}` and getting a "clear" functionality as
default behaviour. If onClearButtonClick is provided, it takes
precendence, so this should be backwards compatible and still allows
customizability of the clear action.