Prevent DataTable from running handleResize after unmount
When the component unmounts, it must clean up the pending timer that it
(might) be running through handleResize. This private method uses
lodash's debounce to run a function on next tick to "debounce"
asynchronous calls. This introduces a bug because the handle function
calls setState which can't be done after a function has unmounted.