Differentiate use of the term "selectable"
Within the IndexProvider the value is accepted as a prop but isn't being
used, when it should be. There's already an existing identifier
"selectable" which can be renamed as "actionable" because it indicates
whether or not actions can be made (are there bulk actions available).
Similarly in the useBulkSelectionData hook there's an existing
identifier "selectable" where it means "does it currently have any
selections. It's a temporary variable created to indicate whether or not
any selections have been made, so we rename it to "hasSelections" even
though the scope doesn't conflict - just to make it clearer.