Maintain backwards compatibility of component with default selectable
This was previously always the case, and when the selectable prop is not
provided we would still expect it to default to true, because the main
behaviour of this component is for use as a selectable index.

This provides backwards compatibility, while still supporting explicitly
passing this value as true/false.