Separate out rendering of headings and of checkboxes
Simplifies the closure function because it no longer has multiple return
signatures (returning either JSX or JSX[], depending on whether or not
its the first element).

Splits out a second one that's specifically for rendering the checkbox,
and then explicitly calling that and pushing it into the array when
selectable.