Remove the esnext folder
This folder was a relic from a time before sewing-kit supported reading
esnext entrypoints from the `esnext` key in package.json, which was
added in sewing-kit 0.135.0 (released Aug 2020).

The configuration that forced sewing-kit to always read from this folder
(instead of looking at the esnext entrypoint) was removed in
https://github.com/Shopify/sewing-kit/pull/2566 and released as
sewing-kit 0.152.0 (released Mar 2021).

If you are building an app using sewing-kit, you must use sewing-kit
0.152.0 or newer if you want esnext build support.