Fix build-consumer.js script
The script was breaking because the `LICENSE.md` file exists in the root dir so the script was breaking. Additionally, changes to docs (i.e., component README.md) weren't being picked up so the root directory was also updated since `polaris-react` now exists a level in from the root folder instead of in the root dir like it originally did.