[Combobox/Listbox] Fix accessibility issues on focus and option create (#5298)
* [Combobox] Fix accessibility issue on focus and option create

When the activator on the Combobox is focused, fix functionality to set active option as first valid option in Listbox. Also fixes bug where popover was closing when creating an option which was inconsist with behavior when selecting/deselecting an existing option.

Co-authored-by: Chloe Rice <chloe.rice@shopify.com>

* [Listbox] Fix logic to reset and find next valid option on create

Fixes logic that wasn't setting create option as active when creating a new option that has no match to existing listbox options. Also fixes logic that wasn't setting next active option if current active option was the first element in listbox when keying up after creating a new option.

Also adds new logic to new reset active option when onScrollToBottom passed in and invoked.

* Update UNRELEASED.md

Co-authored-by: Chloe Rice <chloe.rice@shopify.com>