Delete vocabulary.md
The vocabulary page was originally created when Polaris.Shopify.com launched and our terminology set at Shopify was small. Shopify has grown, and this format doesn't scale for our current needs. At one point, we thought the terms might be useful on our site for 3p developers, but we now have .dev as a resource for these information needs. There is also a project within Shopify around terminology management to solve this problem, plus we have an existing internal spreadsheet that is the most up to date terminology home that crafters use today.