fix: Remove unnecessary use of resetModules
There's no need to reset modules here because the reference to NODE_ENV
in the system under test is within the component and is not at the
module level. It will be evaluated every time at runtime and can be
safely mocked out without needing to reset the modules.

This fix is needed because resetting the modules in this way results in
the components under test being unmounted.