fix: Provide types for useCallback functions
In React 18 types, useCallback wrapped functions require their
parameter types to be explicitly defined. Previously these would end up
being implicit any.

In one case here I've opted to remove the callback altogether because
all it did was wrap a setState with the same value, and that setState
function is already stable so it doesn't require the useCallback
wrapper.