fix: Provide better type narrowing
Previously content was assigned to a type "A | B" and then went into a
conditional block that narrowed it to just "A". But the rendered content
beneath still believed the type to be "A | B" which was allowed
previously where React would treat it as potentially renderable, but it
is now stricter and requires us to do more explicit type narrowing. We
can do that by assigning the value conditionally in the two different
cases so that it knows that it will _always_ be narrowed to "A"