Tokens: Add missing alpha values in RGBA colors (#6975)
This PR adds in some missing alpha values to a couple of our polaris token colors. All tokens are currently using the rbga color format, and there are a couple that didn't define an alpha value. In a browser, a value of 1 will be assumed, so it's not entirely necessary. However, in other implementations like React Native, the alpha value is required or the color will be ignored.

Co-authored-by: Alex Page <hi@alexpage.dev>