[IndexTable] Update headings style and add sortable functionality. (#6949)
WHY are these changes introduced?
Fixes Shopify/web#70689
Fixes Shopify/web#70690
Our IndexTables are sortable across the admin, but there is no way of showing the current sorted state within an IndexTable currently, which could cause confusion for merchants. We also want to tighten up the appearance of the IndexTable to coincide with our efforts to unify the index filtering and searching experience.
WHAT is this pull request doing?

This PR does two things that are related to the IndexTable headings. Firstly, it tightens up the density of the headings by shrinking the vertical padding of a heading, shrinking the font size down to 12px, and also adds in a --p-surface-subdued background color with --p-text-subdued font color to the heading.
Secondly, it adds the ability to show the sorting state of an IndexTable and controlling the sorting state by interacting with a sortable IndexTable heading. The logic introduced here mirrors the logic recently introduced for the DataTable for how that allows for sorting.

How to 🎩
🖥 Local development instructions
🗒 General tophatting guidelines
📄 Changelog guidelines

Clone this branch locally.
Visit the new "With Sortable Headings" story in the IndexTable
Confirm that the Name and Location columns are now sortable through the heading
Confirm that the other columns function as before with no additional sortable functionality

🎩 checklist

 Tested on mobile
 Tested on multiple browsers
 Tested for accessibility
 Updated the component's README.md with documentation changes
 Tophatted documentation changes in the style guide

Co-authored-by: translation-platform[bot] <34770790+translation-platform[bot]@users.noreply.github.com>