[Popover] fix inappropriate closing issue (#7111)
### WHY are these changes introduced?

Fixes #7097  

There are two issues we're trying to address here: 1) popovers will
inappropriately close when a click event removes an element from the
DOM, 2) popovers (and their descendants) will inappropriately close when
a user clicks inside a descendant portal (another popover, modal, etc).

### WHAT is this pull request doing?

To fix the first issue, we propose using `event.composedPath()` to
determine if the event happened inside the popover. This doesn't rely on
the `target` of the event existing in the DOM at the time of the check.

To fix the second issue, we propose checking whether the click event
happened inside another portal by checking all portals within the
`PortalsContainerElement` element, which we can get from
`PortalsManagerContext`. This makes the assumption that any other
portals on the page are a descendant of the original portal. However,
this may not always be the case (e.g. a modal that launches a popover,
in which case the modal is a parent of the popover), so we elected to
make this an opt-in flag.

### 🎩 checklist

I’ve tophatted these changes in the following browsers:

- [x] Chrome latest
- [x] FF latest
- [x] Safari latest
- [x] Edge
- [ ] In at least one of the above browsers, test both retina and
non-retina displays
- [x] iPhone (5/SE/X) (10+) Safari Mobile
- [x] iPad (10+) Safari Mobile
- [x] Android device (5.x) Chrome

- [x] tophatted in web on `main`
- [x] [Tophatted
documentation](https://github.com/Shopify/polaris/blob/main/documentation/Tophatting%20documentation.md)
changes in the style guide