[IndexTable] Fix overflowing sticky column (#7130)
### WHY are these changes introduced?

Fixes #4012 

When a table column is sticky and bigger than the container (_aka_,
causes an overflow) the table width is not calculated properly and the
non-sticky columns overlap the sticky one. Check this [pure html
reproduction](https://codesandbox.io/s/icy-river-0srh0k?file=/index.html)
for a better understanding of the root issue.

Disclaimer: This PR does not fit the _condensed_ mode of the table,
which has it's own issues when overflowing

### WHAT is this pull request doing?

The bug does not happen on small screens because the only sticky column
in that size is the one with checkboxes, which is small enough to not
cause an issue, but when a sticky column has long texts the problem can
happen on bigger screens.

To solve it I calculate dynamically the size of the sticky column(s) and
apply the sticky styles only when the full width of the table can fit
them without an overflow. If the table is narrower than the sticky
columns, the small screen styles are applied (only the checkbox remains
sticky)

If users do not want this behaviour, they can still apply a `max-width`
and `white-space: wrap` to the wider cells or columns, which is the
non-official solution to this bug right now (`web` is doing it, in case
you want to check that the fix does not affect that setup)

  <details>
    <summary>Fixed table (gif)</summary>
<img src="https://screenshot.click/index-table-fix-1.gif" alt="Index
table with two rows, checkboxes and four more columns. The first column
is wider than the table. When scrolling horizontally, only the checkbox
remains sticky and the rest of the content scrolls">
  </details>


### 🎩 checklist

[Spin instance
](https://admin.web.web-kf1h.paqui-calabria.eu.spin.dev/store/shop1/)
- [ ] Tested on
[mobile](https://github.com/Shopify/polaris/blob/main/documentation/Tophatting.md#cross-browser-testing)
- [ ] Tested on [multiple
browsers](https://help.shopify.com/en/manual/shopify-admin/supported-browsers)
- [ ] Tested for
[accessibility](https://github.com/Shopify/polaris/blob/main/documentation/Accessibility%20testing.md)
- [ ] Updated the component's `README.md` with documentation changes
- [ ] [Tophatted
documentation](https://github.com/Shopify/polaris/blob/main/documentation/Tophatting%20documentation.md)
changes in the style guide