Add documentation on breakpoint token usage in media queries (#7230)
<!--
  ☝️How to write a good PR title:
- Prefix it with [ComponentName] (if applicable), for example: [Button]
  - Start with a verb, for example: Add, Delete, Improve, Fix…
  - Give as much context as necessary and as little as possible
  - Prefix it with [WIP] while it’s a work in progress
-->

### WHY are these changes introduced?

We've noticed some incorrect implementations of the new breakpoint
tokens and want to add more clarifying documentation to the tokens page.

The breakpoint tokens are a bit special in that a transform takes the
values and generates Sass variables for each breakpoint in `up`, `down`,
and `only` directions. These Sass variables are the values that should
be used in media queries not the tokens themselves.

<!--
  Context about the problem that’s being addressed.
-->

### WHAT is this pull request doing?

<!--
  Summary of the changes committed.

Before / after screenshots are appreciated for UI changes. Make sure to
include alt text that describes the screenshot.

If you include an animated gif showing your change, wrapping it in a
details tag is recommended. Gifs usually autoplay, which can cause
accessibility issues for people reviewing your PR:

    <details>
      <summary>Summary of your gif(s)</summary>
      <img src="..." alt="Description of what the gif shows">
    </details>
-->

This PR adds descriptions to each of breakpoint tokens as well as a
section below the tokens to go into further detail about how to use them
in media queries. It's not the most elegant section but it contains
really important information we need tied to the breakpoint tokens.
Happy to jam on tweaking the styling for this section further if needed.


![localhost_3000_tokens_breakpoints](https://user-images.githubusercontent.com/21976492/191132212-6a82af0a-3408-4e35-9269-b6b6bfc88190.png)