[polaris.shopify.com] Update IA and navigation (#7232)
This PR Introduces a new IA for the Polaris website 🚀

Closes #7294
Closes #6417
Closes #6657
Closes #6705


https://user-images.githubusercontent.com/875708/192775921-5d26b714-fe19-465a-97f2-6d6fc8b448ec.mp4

This PR adds support for the new IA that JJ has championed. To
accommodate the new IA, I had to rip out a ton of stuff and touch up the
design.

The PR is quite big — sorry about that — but most of it is moved
markdown files and deleted stuff. On the whole, this PR reduces the
total lines of code.

✅ Added a new side nav, built on @gui-santos awesome work on the mobile
nav.
✅ Removed `MobileNav`. It's part of the Frame component now.
✅ Combined `FoundationsGrid` and `ComponentsGrid` into `Grid`.
✅ Renamed `Layout` to `Page` to clarify what it does
✅ Removed a bunch of redundant `getStaticProps` by using a catch-all
template that automatically deals with most pages.
✅ Simplified the script that generates nav.json. Removed all hard coded
nav items.

Co-authored-by: jjgali <77791660+jjgali@users.noreply.github.com>
Co-authored-by: Gui Santos <guilherme.santos@shopify.com>