[Page] Use state callback in measure actions (#7361)
Partial fix for https://github.com/Shopify/polaris/issues/7330

[Spin link](https://admin.web.web-1mvc.kyle-durand.us.spin.dev/store/)

Part two might be a css based fix. The heavy re-rendering in the admin
introduced with the React 18 work is causing a ton of layout thrashing
in this component. The root cause will definitely need to be fixed but
we'll do what we can on our end in the meantime

Co-authored-by: Henry Yi <jita.yi@shopify.com>