refactor: rewrite Scrollable utilizing platform features
This updates the Scrollable component to be a function component, but
also changes the fundamental implementation strategy.

Previously, the Scrollable component held internal state for the scroll
position and controlled the scroll of the actual DOM node as a side
effect.

With this change, there is no controlling of the DOM node's scroll,
allowing the platform to handle all scrolling interaction and only using
native features to supply enhancements for the ScrollTo component, the
hint functionality.