Simplify design patterns page (#7441)
### WHY are these changes introduced?

The design pattern page currently includes a section describing the
differences between compositions and patterns. While this information is
valuable to us a team, we are still defining and testing these concepts.
Publishing them on the site at this stage seems to have introduced some
confusion for readers.

### WHAT is this pull request doing?

Removes section on compositions vs patterns until we develop stronger
opinions and feel comfortable publishing them publicly.

Also reorganizes the order of the principles, placing "Contextual" last,
and closest to the related image.