Add migration for sass z-index fn (#7310)
WHY are these changes introduced?
Part of #7211 

WHAT is this pull request doing?
Adding a replace-sass-z-index migration script.

Running npx @shopify/polaris-migrator replace-sass-z-index "**/*.scss"
will target z-index function usage and do the following:

* If more than one argument exists we assume the code is invoking
z-index to map-get on a custom sass map. In this case, we add a comment
with a migration warning and code snippet replacing the
`legacy-polaris-v8.z-index` usage with `map-get`
* If there's a calculation in the same declaration but the
`legacy-polaris-v8.z-index` invocation is valid and corresponds to a
polaris token, we add a comment with a migration warning and a code
snippet replacing that invocation with a css variable invocation
pointing at the corresponding polaris-token.
* If its just a valid `legacy-polaris-v8.z-index` invocation that
corresponds to an existing polaris token, we replace it with a
css-variable invocation pointing at the corresponding polaris-token.
* Otherwise we ignore the declaration. 

Things our codemod doesn't presently understand but should: 
* interpolations. `calc(#{legacy-polaris-v8.z-index} + 1)` for example
is completely ignored by our code. As post-css interprets this as an
invocation of a function called `#{legacy-polaris-v8.z-index`. It
doesn't seem to understand interpolations at the moment. We can probably
use something like style-lint's
[hasInterpolation](https://github.dev/stylelint/stylelint/blob/eee9deb35190f46c5f964769d521fef5d7d1d7a8/lib/utils/__tests__/hasInterpolation.test.js#L5)
to work around this, (ty @aaronccasanova), but not something we should
worry about for first pass imo.

Co-authored-by: Aaron Casanova <32409546+aaronccasanova@users.noreply.github.com>