Always display subtitle in Page
If a <Page> component only has subtitle has a property,
the `hasHeaderContent` check will return false and no header will be rendered.

With this commit we add the possiblility to render a header inside a Page
if it only has `subtitle` set as a prop.
We also amend the test to reflect that.