[ResourceItem] Use layout components (#7784)
Closes: https://github.com/Shopify/polaris/issues/7580

This slightly moves the checkbox handle 2px to the left. This means it
now lines up with our 20px padding instead of 22px from the card edge.
Also improves vertical alignment of the checkboxes and bulk select on xs
screens

Before | After
---|---
![image](https://screenshot.click/24-12-z59dz-bpdqu.png) |
![image](https://screenshot.click/24-12-viiqs-efiaf.png)

The only CSS that remains is for interactive elements, interactive
behavior (hover, focus), and transitions (which are being removed in
https://github.com/Shopify/polaris/pull/7290/files).