[BulkActions] Improve visibility of the BulkActions bar (#7851)
### WHY are these changes introduced?

Some feedback we got from SLT around the new Bulk Actions bar UX is
that, whilst a nice update, it can sometimes be hard to detect that the
bulk actions bar has appeared. This PR addresses that by tweaking some
of the settings for the sticky Bulk Actions bar.

### WHAT is this pull request doing?

This PR updates the bulk actions to use a "dark UI" pattern, similar to
the ContextualSaveBar and Toast components currently in Polaris. The
placement of the bar has been updated to be a little higher on the
screen, up from 16px from the bottom, to 48px from the bottom. We have
also combined the `--p-shadow-deep` and `--p-shadow-popover` shadow
tokens to increase the visual lift the component has from the rest of
the index UI.



https://user-images.githubusercontent.com/2562596/206267671-6782a7bc-f9c6-47ff-8a04-da374fcdf499.mov



### 🎩 checklist

- [x] Tested on
[mobile](https://github.com/Shopify/polaris/blob/main/documentation/Tophatting.md#cross-browser-testing)
- [x] Tested on [multiple
browsers](https://help.shopify.com/en/manual/shopify-admin/supported-browsers)
- [x] Tested for
[accessibility](https://github.com/Shopify/polaris/blob/main/documentation/Accessibility%20testing.md)
- [x] Updated the component's `README.md` with documentation changes
- [x] [Tophatted
documentation](https://github.com/Shopify/polaris/blob/main/documentation/Tophatting%20documentation.md)
changes in the style guide

Co-authored-by: Lo Kim <lo.kim@shopify.com>