[Layout] Rebuild `OptionList` with layout primitives (#7958)
### WHY are these changes introduced?

Fixes #7952 

### WHAT is this pull request doing?

Updates `OptionList` to use new layout primitives.

One small visual change with this update due to replacing `margin` with
`padding` is a slight increase (`4px`) in spacing when there are
multiple sections in an `OptionList`. I don't think the added complexity
to keep this the same as current doesn't seem worth it, but happy to
hear other thoughts.

**Update:** Added a `Bleed` to resolve the spacing change. Feels hacky
but prevents the visual change.