Remove downlevel-dts and legacy type support (#8203)
### WHY are these changes introduced?

Polaris currently supports multiple TypeScript versions by building our
types twice. This breaking change in v11 removes this functionality and
we will only support the current version in the library.

This used to be required to build the props table on
legacy.polaris.shopify.com. This is no longer a requirement with
@martenbjork brilliant work.

### WHAT is this pull request doing?

Removes legacy type support for the library.

---------

Co-authored-by: Ben Scott <227292+BPScott@users.noreply.github.com>
Co-authored-by: Aaron Casanova <32409546+aaronccasanova@users.noreply.github.com>