Add new z-index token names (#8245)
<!--
  ☝️How to write a good PR title:
- Prefix it with [ComponentName] (if applicable), for example: [Button]
  - Start with a verb, for example: Add, Delete, Improve, Fix…
  - Give as much context as necessary and as little as possible
  - Prefix it with [WIP] while it’s a work in progress
-->

### WHY are these changes introduced?

We want to rename these tokens in v11. By adding the new names in v10 we
can break up the major version bump into more manageable chunks (we can
create and run migrations for the values in v10 and then just remove the
values in the launch of v11).

<!--
  Context about the problem that’s being addressed.
-->

### WHAT is this pull request doing?

<!--
  Summary of the changes committed.

Before / after screenshots are appreciated for UI changes. Make sure to
include alt text that describes the screenshot.

If you include an animated gif showing your change, wrapping it in a
details tag is recommended. Gifs usually autoplay, which can cause
accessibility issues for people reviewing your PR:

    <details>
      <summary>Summary of your gif(s)</summary>
      <img src="..." alt="Description of what the gif shows">
    </details>
-->

<!-- ℹ️ Delete the following for small / trivial changes -->

Adds a set of identical z-index tokens with the names `--p-z-index-*`
(vs. `--p-z-*`)

---------

Co-authored-by: Alex Page <hi@alexpage.dev>
Co-authored-by: Aaron Casanova <32409546+aaronccasanova@users.noreply.github.com>