Content edits
Since we aren't including the Figma file link right now, the designing section was pretty sparse. I brought `Breakpoints and spacing` up to an h2, and moved the line about browsing components to the top section of the page.

We can consider restructuring this as more assets like the Figma file become available.