Tooltip: Change easing function to `ease-in` (#8839)
Changes the easing of tooltips from `ease-in` to `ease-out`.

IMO, the using `ease-in` causes the tooltip to "slam" into its final
position. Replacing it with `ease-out` makes for a softer entry
animation. Hardly noticeable with the naked eye, but it feels better.

---

Before: (slowed down, 50% speed)


https://user-images.githubusercontent.com/875708/229754881-58959ea9-a463-4fa7-a465-f66d289f723e.mp4

---

After: (slowed down, 50% speed)


https://user-images.githubusercontent.com/875708/229754932-0bed8fa6-d479-4b2a-ab5a-a90e07c840d7.mp4