Don't setting state during render to support Suspense
When rendering an PortalManager (by means of an AppProvider) with a
component tree that includes a Suspense node we get an error that
Suspense received an update before it finished hydrating.

This comes from the fact that we're using a state setting function as a
ref which causes a synchronous re-render during the render pass of the
component. In order to avoid this, we can use a ref along with a helper
to memoize the value to the component's mounted state. When this
component has finished mounting, it will update some internal state
which will then allow it to re-calculate the context value and pass down
the reference to the now-mounted node.

We still have the duplicate render, but this seems to be avoidable when
it comes to targetting a DOM node which must have first been rendered.