Update `Avatar` (#9380)
### WHY are these changes introduced?

Resolves https://github.com/Shopify/polaris-summer-editions/issues/102


### WHAT is this pull request doing?

- Updates Avatar shape to only be square
- Updates Avatar colors
- Updates Avatar border-radius
- Updates Avatar XS size
- ~Updates Avatar font-sizes~*
- Adds experimental color tokens for decorative colors
- Updates stories to better consolidate examples

_*Right now font-size in the Avatars is scaled based on the medium
avatar font-size. Updating this size to the new mocks would overflow the
backgrounds currently. And it would be too involved to reconfigure the
component to have non-scaled font-sizes for the current scope of work.
If it ends up looking off in other components we can definitely tweak
this value. When we rework/revisit the Avatar in the future we can
explore having non-scaled font-sizes._

### How to 🎩

Review on
[Storybook](https://5d559397bae39100201eedc1-pwakwmvjdb.chromatic.com/?path=/story/all-components-avatar--all&globals=polarisSummerEditions2023:true)
- Ensure beta flagged Avatar styles render as expected
- Ensure no regressions to the existing Avatar styles

---------

Co-authored-by: Aaron Casanova <32409546+aaronccasanova@users.noreply.github.com>