[TopBar] Remove background-color transition and fix search input on Safari (#9463)
### WHY are these changes introduced?

Fixes for issues detected during the implementation of se23 in the
Admin.

### WHAT is this pull request doing?

This PR removes the transition effect of the `background-color` when
loading the `TopBar` with the se23 feature enabled. The transition
effect caused a strange experience when loading the Admin, as the
background color shifted from light to dark.

This PR also addresses an issue detected on Safari. The property
`appearance` was adding a lot of `border-radius` to the input element.
The `appearance` property doesn't seem to be necessary as well since all
the native input styles seem to be overridden correctly already.


https://github.com/Shopify/polaris/assets/2091116/ad16d94a-7654-46e2-9f98-f96e02479132

On the video above it's possible to see the border-radius being applied
when the `appearance` is set to `textfield` on Safari Mobile.

### 🎩 checklist

- [x] Tested on
[mobile](https://github.com/Shopify/polaris/blob/main/documentation/Tophatting.md#cross-browser-testing)
- [x] Tested on [multiple
browsers](https://help.shopify.com/en/manual/shopify-admin/supported-browsers)
- [x] Tested for
[accessibility](https://github.com/Shopify/polaris/blob/main/documentation/Accessibility%20testing.md)
- [ ] Updated the component's `README.md` with documentation changes
- [ ] [Tophatted
documentation](https://github.com/Shopify/polaris/blob/main/documentation/Tophatting%20documentation.md)
changes in the style guide

---------

Co-authored-by: Kyle Durand <kyledurand@users.noreply.github.com>