Replace `AppProvider` `defaultProps` with `getFeatures` method (#9769)
This PR replaces the previous `AppProvider` `static defaultProps` with a
more reliable `getFeatures` method. The issue is that `defaultProps`
does not initialize flags to `false` when a `features` object is
provided with `undefined` values. Consequently, beta-flags were set to
`undefined` and passed to the `force` argument of `classList.toggle`
which toggled class names on each render. By introducing `getFeatures`,
we ensure that features are consistently retrieved with their default
values set to `false`.