ActionsList filtering (#9907)
Closes https://github.com/Shopify/polaris/issues/8686

## Why?
With extension GA comes a lot of potential new items in the `More
Actions` of pages supporting extensibility making it increasingly more
difficult for merchants to find what they are looking for.
The median number of links in "More action" menus on the products,
orders, and customers pages for high GMV merchants is already between 12
and 31. This number does not count any non-app actions that have been
added, so in reality, the number of actions is even higher. This is a
terrible UX for our merchants.

## Proposed Changes

Adding a search input that will appear anytime an ActionList has more
than 10 elements.

<img width="1015" alt="Screenshot 2023-08-09 at 9 06 13 AM"
src="https://github.com/Shopify/ui-api-design/assets/3555826/cb021493-ada4-4ac2-8991-42a6d979207a">
<img width="1261" alt="Screenshot 2023-08-10 at 1 52 48 PM"
src="https://github.com/Shopify/polaris/assets/3555826/6c2d9763-6b81-4808-9659-eb1db7b4bf1c">


This will give us the filtering behaviour in various areas and cases for
free:
- When actionGroups are rolledUp
- In Bulk Actions
- When actionGroups are not rolledUp, but still have more than 10 items

This will also enable the rest of the admin to leverage this behaviour
for free to give a consistent merchant experience.

## Open questions

Using the TextField (or SearchField from filters) creates a circular
dependency with the ActionList. What would be the right approach to have
the search input with an icon without creating circular dependencies?

## Resources

- Action Extensions tech doc -
https://docs.google.com/document/d/1BEAHpkfeyUTaxW3B18LjojimDmqNLr1Fc2wfpMXzQ64/edit
- Mockups for Action Extensions -
https://www.figma.com/proto/afrUykUTB0n4GB4evE5F6R/%F0%9F%92%8EAdmin-UI-extension-patterns?node-id=398-111475&scaling=scale-down&page-id=295%3A118673&starting-point-node-id=398%3A111475&show-proto-sidebar=1
- Post in Admin UX (WxM) with an initial proposal -
https://shopify.workplace.com/groups/1294395484626415/permalink/1321299791935984/
- Investigation of changes needed -
https://github.com/Shopify/app-ui/issues/36#issuecomment-1468527331

---------

Co-authored-by: translation-platform[bot] <34770790+translation-platform[bot]@users.noreply.github.com>